[AIP-21] Move contrib.hooks.azure_fileshare_hook airflow.contrib.hooks.azure_fileshare_hook
