+++ scripts/ci/pylint_todo.txt
@@ -7,7 +7,7 @@
 ./airflow/providers/microsoft/azure/hooks/azure_container_volume.py
 ./airflow/providers/microsoft/azure/hooks/azure_cosmos.py
 ./airflow/providers/microsoft/azure/hooks/azure_data_lake.py
-./airflow/contrib/hooks/azure_fileshare_hook.py
+./airflow/providers/microsoft/azure/hooks/azure_fileshare.py
 ./airflow/contrib/hooks/datadog_hook.py
 ./airflow/contrib/hooks/dingding_hook.py
 ./airflow/contrib/hooks/discord_webhook_hook.py
